backend/pyproject.toml
[project]
name = "disp-time-backend"
version = "0.1.0"
description = "Backend for Display Time Web App"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "sqlmodel",
    "httpx",
    "PyYAML",
    "python-multipart",
]
